server/requirements.txt
PyYAML==6.0.2
anthropic==0.97.0
